+++ aws_codeseeder/services/cfn.py
@@ -18,6 +18,7 @@ from datetime import datetime
 from typing import Any, Dict, Optional, Tuple, cast
 
 import botocore.exceptions
+from boto3 import Session
 
 from aws_codeseeder import LOGGER
 from aws_codeseeder.services import s3
@@ -26,8 +27,8 @@ from aws_codeseeder.services._utils import boto3_client
 CHANGESET_PREFIX = "aws-codeseeder-"
 
 
-def _wait_for_changeset(changeset_id: str, stack_name: str) -> bool:
-    waiter = boto3_client("cloudformation").get_waiter("change_set_create_complete")
+def _wait_for_changeset(changeset_id: str, stack_name: str, session: Optional[Session] = None) -> bool:
+    waiter = boto3_client("cloudformation", session=session).get_waiter("change_set_create_complete")
     waiter_config = {"Delay": 1}
     try:
         waiter.wait(ChangeSetName=changeset_id, StackName=stack_name, WaiterConfig=waiter_config)
@@ -52,11 +53,12 @@ def _create_changeset(
     seedkit_tag: Optional[str] = None,
     template_path: str = "",
     parameters: Optional[Dict[str, str]] = None,
+    session: Optional[Session] = None,
 ) -> Tuple[str, str]:
     now = datetime.utcnow().isoformat()
     description = f"Created by AWS CodeSeeder CLI at {now} UTC"
     changeset_name = CHANGESET_PREFIX + str(int(time.time()))
-    stack_exist, _ = does_stack_exist(stack_name=stack_name)
+    stack_exist, _ = does_stack_exist(stack_name=stack_name, session=session)
     changeset_type = "UPDATE" if stack_exist else "CREATE"
     kwargs: Dict[str, Any] = {
         "ChangeSetName": changeset_name,
@@ -74,19 +76,19 @@ def _create_changeset(
         kwargs.update({"TemplateURL": template_path})
     if parameters:
         kwargs.update({"Parameters": [{"ParameterKey": k, "ParameterValue": v} for k, v in parameters.items()]})
-    resp = boto3_client("cloudformation").create_change_set(**kwargs)
+    resp = boto3_client("cloudformation", session=session).create_change_set(**kwargs)
     return str(resp["Id"]), changeset_type
 
 
-def _execute_changeset(changeset_id: str, stack_name: str) -> None:
-    boto3_client("cloudformation").execute_change_set(ChangeSetName=changeset_id, StackName=stack_name)
+def _execute_changeset(changeset_id: str, stack_name: str, session: Optional[Session] = None) -> None:
+    boto3_client("cloudformation", session=session).execute_change_set(ChangeSetName=changeset_id, StackName=stack_name)
 
 
-def _wait_for_execute(stack_name: str, changeset_type: str) -> None:
+def _wait_for_execute(stack_name: str, changeset_type: str, session: Optional[Session] = None) -> None:
     if changeset_type == "CREATE":
-        waiter = boto3_client("cloudformation").get_waiter("stack_create_complete")
+        waiter = boto3_client("cloudformation", session=session).get_waiter("stack_create_complete")
     elif changeset_type == "UPDATE":
-        waiter = boto3_client("cloudformation").get_waiter("stack_update_complete")
+        waiter = boto3_client("cloudformation", session=session).get_waiter("stack_update_complete")
     else:
         raise RuntimeError(f"Invalid changeset type {changeset_type}")
     waiter_config = {
@@ -112,13 +114,15 @@ def get_stack_name(seedkit_name: str) -> str:
     return f"aws-codeseeder-{seedkit_name}"
 
 
-def get_stack_status(stack_name: str) -> str:
+def get_stack_status(stack_name: str, session: Optional[Session] = None) -> str:
     """Retrieve the status of a CloudFormation Stack
 
     Parameters
     ----------
     stack_name : str
         Name of the CloudFormation Stack to query
+    session: Optional[Session], optional
+        Optional Session to use for all boto3 operations, by default None
 
     Returns
     -------
@@ -131,7 +135,7 @@ def get_stack_status(stack_name: str) -> str:
     ValueError
         If the Stack is not found
     """
-    client = boto3_client("cloudformation")
+    client = boto3_client("cloudformation", session=session)
     try:
         resp = client.describe_stacks(StackName=stack_name)
         if len(resp["Stacks"]) < 1:
@@ -141,20 +145,22 @@ def get_stack_status(stack_name: str) -> str:
     return cast(str, resp["Stacks"][0]["StackStatus"])
 
 
-def does_stack_exist(stack_name: str) -> Tuple[bool, Dict[str, str]]:
+def does_stack_exist(stack_name: str, session: Optional[Session] = None) -> Tuple[bool, Dict[str, str]]:
     """Checks for existence of a CloudFormation Stack while also returning Stack Outputs if it does exist
 
     Parameters
     ----------
     stack_name : str
         Name of the CloudFormation Stack to query
+    session: Optional[Session], optional
+        Optional Session to use for all boto3 operations, by default None
 
     Returns
     -------
     Tuple[bool, Dict[str, str]]
         Tuple2 with a boolean indicating Stack existence and a dict of any Stack Outputs
     """
-    client = boto3_client("cloudformation")
+    client = boto3_client("cloudformation", session=session)
     try:
         resp = client.describe_stacks(StackName=stack_name)
         if len(resp["Stacks"]) < 1:
@@ -174,6 +180,7 @@ def deploy_template(
     seedkit_tag: Optional[str] = None,
     s3_bucket: Optional[str] = None,
     parameters: Optional[Dict[str, str]] = None,
+    session: Optional[Session] = None,
 ) -> None:
     """Deploy a local CloudFormation Template
 
@@ -192,6 +199,8 @@ def deploy_template(
         S3 Bucket to upload the template file to if it is too large (> 51200), by default None
     parameters: Optional[Dict[str, str]], optional
         Key/Value set of Input Parameters to pass to the CloudFormation stack, by default None
+    session: Optional[Session], optional
+        Optional Session to use for all boto3 operations, by default None
 
     Raises
     ------
@@ -213,32 +222,42 @@ def deploy_template(
         key = f"cli/remote/demo/{s3_file_name}"
         s3_template_path = f"https://s3.amazonaws.com/{s3_bucket}/{key}"
         LOGGER.debug("s3_template_path: %s", s3_template_path)
-        s3.upload_file(src=local_template_path, bucket=s3_bucket, key=key)
+        s3.upload_file(src=local_template_path, bucket=s3_bucket, key=key, session=session)
         changeset_id, changeset_type = _create_changeset(
-            stack_name=stack_name, template_str="", seedkit_tag=seedkit_tag, template_path=s3_template_path
+            stack_name=stack_name,
+            template_str="",
+            seedkit_tag=seedkit_tag,
+            template_path=s3_template_path,
+            session=session,
         )
     else:
         with open(filename, "r") as handle:
             template_str = handle.read()
         changeset_id, changeset_type = _create_changeset(
-            stack_name=stack_name, template_str=template_str, seedkit_tag=seedkit_tag, parameters=parameters
+            stack_name=stack_name,
+            template_str=template_str,
+            seedkit_tag=seedkit_tag,
+            parameters=parameters,
+            session=session,
         )
-    has_changes = _wait_for_changeset(changeset_id, stack_name)
+    has_changes = _wait_for_changeset(changeset_id, stack_name, session=session)
     if has_changes:
-        _execute_changeset(changeset_id=changeset_id, stack_name=stack_name)
-        _wait_for_execute(stack_name=stack_name, changeset_type=changeset_type)
+        _execute_changeset(changeset_id=changeset_id, stack_name=stack_name, session=session)
+        _wait_for_execute(stack_name=stack_name, changeset_type=changeset_type, session=session)
 
 
-def destroy_stack(stack_name: str) -> None:
+def destroy_stack(stack_name: str, session: Optional[Session] = None) -> None:
     """Destroy the CloudFormation Stack
 
     Parameters
     ----------
     stack_name : str
         Name of the CloudFormation Stack to destroy
+    session: Optional[Session], optional
+        Optional Session to use for all boto3 operations, by default None
     """
     LOGGER.debug("Destroying stack %s", stack_name)
-    client = boto3_client("cloudformation")
+    client = boto3_client("cloudformation", session=session)
     client.delete_stack(StackName=stack_name)
     waiter = client.get_waiter("stack_delete_complete")
     waiter.wait(StackName=stack_name, WaiterConfig={"Delay": 5, "MaxAttempts": 200})